Refactor component detectors into static classes
