+++ lib/solvers/ComponentDetectionSolver/RectBoundsComponentDetectionStage.ts
@@ -4,12 +4,12 @@ import type { GraphicsObject } from "graphics-debug"
 import { getStringColor, safeTransparentize } from "lib/solvers/colors"
 import type { Obstacle, SimpleRouteJson } from "lib/types"
 import { getBoundsForObstacles } from "lib/utils/getBoundsForObstacles"
-import { getViaDimensions } from "lib/utils/getViaDimensions"
 import type {
   ComponentDetectionSolverOutput,
   ComponentDetectionSolverParams,
   DetectedComponent,
 } from "./ComponentDetectionSolver"
+import { detectComponentKind, type ComponentKind } from "./detectors"
 
 const cloneObstacle = (obstacle: Obstacle): Obstacle => ({ ...obstacle })
 const cloneObstacles = (obstacles: Obstacle[]) => obstacles.map(cloneObstacle)
@@ -35,355 +35,6 @@ const doObstaclesOverlap = ({
   replacementObstacle: Obstacle
 }) =>
   doBoundsOverlap(getBoundingBox(obstacle), getBoundingBox(replacementObstacle))
-const MIN_BGA_AXIS_COUNT = 3
-const MIN_BGA_TWO_AXIS_COUNT = 2
-const MIN_BGA_LONG_AXIS_COUNT_FOR_TWO_AXIS = 4
-const MIN_BGA_GRID_OCCUPANCY = 0.5
-const MAX_BGA_PAD_SIZE_VARIANCE = 0.01
-const AXIS_CLUSTER_EPSILON = 1e-3
-const MIN_QFP_PADS_PER_SIDE = 3
-const MAX_QFP_PAD_COUNT = 32
-const MAX_QFP_THERMAL_PAD_COUNT = 4
-const MAX_QFP_CENTER_NEAREST_SIDE_RATIO = 0.25
-const MIN_QFP_PAD_ASPECT_RATIO = 1.5
-
-type ComponentKind = DetectedComponent["componentKind"]
-
-function clusterAxisValues(values: number[]) {
-  const sortedValues = [...values].sort((a, b) => a - b)
-  const clustered: number[] = []
-
-  for (const value of sortedValues) {
-    const previousValue = clustered[clustered.length - 1]
-    if (
-      previousValue === undefined ||
-      Math.abs(value - previousValue) > AXIS_CLUSTER_EPSILON
-    ) {
-      clustered.push(value)
-    }
-  }
-
-  return clustered
-}
-
-function hasUniformDimensionWithinTolerance(values: number[]) {
-  const minValue = Math.min(...values)
-  const maxValue = Math.max(...values)
-
-  if (minValue <= 0) return false
-
-  return maxValue / minValue <= 1 + MAX_BGA_PAD_SIZE_VARIANCE
-}
-
-function hasUniformPadDimensions(memberObstacles: Obstacle[]) {
-  return (
-    hasUniformDimensionWithinTolerance(
-      memberObstacles.map((obstacle) => obstacle.width),
-    ) &&
-    hasUniformDimensionWithinTolerance(
-      memberObstacles.map((obstacle) => obstacle.height),
-    )
-  )
-}
-
-function isBgaLikeComponent(memberObstacles: Obstacle[]) {
-  if (!hasUniformPadDimensions(memberObstacles)) return false
-
-  const rowAxisValues = clusterAxisValues(
-    memberObstacles.map((obstacle) => obstacle.center.y),
-  )
-  const columnAxisValues = clusterAxisValues(
-    memberObstacles.map((obstacle) => obstacle.center.x),
-  )
-  const rowCount = rowAxisValues.length
-  const columnCount = columnAxisValues.length
-  const gridCellCount = rowCount * columnCount
-  const gridOccupancy =
-    gridCellCount > 0 ? memberObstacles.length / gridCellCount : 0
-  const hasStandardBgaAxisCounts =
-    rowCount >= MIN_BGA_AXIS_COUNT && columnCount >= MIN_BGA_AXIS_COUNT
-  const hasTwoAxisBgaAxisCounts =
-    (rowCount === MIN_BGA_TWO_AXIS_COUNT &&
-      columnCount >= MIN_BGA_LONG_AXIS_COUNT_FOR_TWO_AXIS) ||
-    (columnCount === MIN_BGA_TWO_AXIS_COUNT &&
-      rowCount >= MIN_BGA_LONG_AXIS_COUNT_FOR_TWO_AXIS)
-  const hasSupportedAxisCounts =
-    hasStandardBgaAxisCounts || hasTwoAxisBgaAxisCounts
-
-  return hasSupportedAxisCounts && gridOccupancy >= MIN_BGA_GRID_OCCUPANCY
-}
-
-function getNearestClusterIndex(value: number, clusters: number[]) {
-  let nearestIndex = 0
-  let nearestDistance = Number.POSITIVE_INFINITY
-
-  for (let index = 0; index < clusters.length; index++) {
-    const distance = Math.abs(value - clusters[index]!)
-    if (distance < nearestDistance) {
-      nearestIndex = index
-      nearestDistance = distance
-    }
-  }
-
-  return nearestIndex
-}
-
-function getTwoAxisInnerGap({
-  memberObstacles,
-  rowAxisValues,
-  columnAxisValues,
-}: {
-  memberObstacles: Obstacle[]
-  rowAxisValues: number[]
-  columnAxisValues: number[]
-}) {
-  if (rowAxisValues.length === 2 && columnAxisValues.length >= 4) {
-    const rows = [[], []] as [Obstacle[], Obstacle[]]
-
-    for (const obstacle of memberObstacles) {
-      rows[getNearestClusterIndex(obstacle.center.y, rowAxisValues)]!.push(
-        obstacle,
-      )
-    }
-
-    return (
-      Math.min(
-        ...rows[1].map((obstacle) => obstacle.center.y - obstacle.height / 2),
-      ) -
-      Math.max(
-        ...rows[0].map((obstacle) => obstacle.center.y + obstacle.height / 2),
-      )
-    )
-  }
-
-  if (columnAxisValues.length === 2 && rowAxisValues.length >= 4) {
-    const columns = [[], []] as [Obstacle[], Obstacle[]]
-
-    for (const obstacle of memberObstacles) {
-      columns[
-        getNearestClusterIndex(obstacle.center.x, columnAxisValues)
-      ]!.push(obstacle)
-    }
-
-    return (
-      Math.min(
-        ...columns[1].map((obstacle) => obstacle.center.x - obstacle.width / 2),
-      ) -
-      Math.max(
-        ...columns[0].map((obstacle) => obstacle.center.x + obstacle.width / 2),
-      )
-    )
-  }
-
-  return null
-}
-
-function isSoicLikeComponent({
-  memberObstacles,
-  inputSrj,
-}: {
-  memberObstacles: Obstacle[]
-  inputSrj: SimpleRouteJson
-}) {
-  if (!isBgaLikeComponent(memberObstacles)) return false
-
-  const rowAxisValues = clusterAxisValues(
-    memberObstacles.map((obstacle) => obstacle.center.y),
-  )
-  const columnAxisValues = clusterAxisValues(
-    memberObstacles.map((obstacle) => obstacle.center.x),
-  )
-  const isTwoRowOrColumnBga =
-    (rowAxisValues.length === 2 && columnAxisValues.length >= 4) ||
-    (columnAxisValues.length === 2 && rowAxisValues.length >= 4)
-
-  if (!isTwoRowOrColumnBga) return false
-
-  const innerGap = getTwoAxisInnerGap({
-    memberObstacles,
-    rowAxisValues,
-    columnAxisValues,
-  })
-  if (innerGap === null) return false
-
-  const viaDiameter = getViaDimensions(inputSrj).padDiameter
-  const obstacleMargin = inputSrj.defaultObstacleMargin ?? 0.15
-  const soicGapThreshold = (viaDiameter + obstacleMargin) * 2
-
-  return innerGap > soicGapThreshold
-}
-
-function getNearestSideCounts(memberObstacles: Obstacle[]) {
-  const bounds = getBoundsForObstacles(memberObstacles)
-  const width = bounds.maxX - bounds.minX
-  const height = bounds.maxY - bounds.minY
-  const counts = {
-    top: 0,
-    right: 0,
-    bottom: 0,
-    left: 0,
-  }
-
-  if (width <= 0 || height <= 0) return { counts, maxNearestSideRatio: 1 }
-
-  let maxNearestSideRatio = 0
-
-  for (const obstacle of memberObstacles) {
-    const isHorizontalPad = obstacle.width > obstacle.height
-    const distances = [
-      {
-        side: "top" as const,
-        distance: Math.abs(obstacle.center.y - bounds.minY),
-        axisSpan: height,
-        orientationMatches: !isHorizontalPad,
-      },
-      {
-        side: "right" as const,
-        distance: Math.abs(bounds.maxX - obstacle.center.x),
-        axisSpan: width,
-        orientationMatches: isHorizontalPad,
-      },
-      {
-        side: "bottom" as const,
-        distance: Math.abs(bounds.maxY - obstacle.center.y),
-        axisSpan: height,
-        orientationMatches: !isHorizontalPad,
-      },
-      {
-        side: "left" as const,
-        distance: Math.abs(obstacle.center.x - bounds.minX),
-        axisSpan: width,
-        orientationMatches: isHorizontalPad,
-      },
-    ]
-      .filter((candidate) => candidate.orientationMatches)
-      .sort((a, b) => a.distance - b.distance)
-    const nearest = distances[0]!
-
-    counts[nearest.side] += 1
-    maxNearestSideRatio = Math.max(
-      maxNearestSideRatio,
-      nearest.distance / nearest.axisSpan,
-    )
-  }
-
-  return { counts, maxNearestSideRatio }
-}
-
-function isQfpPerimeterPadObstacle(obstacle: Obstacle) {
-  const minSide = Math.min(obstacle.width, obstacle.height)
-  const maxSide = Math.max(obstacle.width, obstacle.height)
-
-  return minSide > 0 && maxSide / minSide >= MIN_QFP_PAD_ASPECT_RATIO
-}
-
-function getQfpPerimeterPadObstacles(memberObstacles: Obstacle[]) {
-  return memberObstacles.filter(isQfpPerimeterPadObstacle)
-}
-
-function areCentralThermalPadObstacles({
-  perimeterPadObstacles,
-  thermalPadObstacles,
-}: {
-  perimeterPadObstacles: Obstacle[]
-  thermalPadObstacles: Obstacle[]
-}) {
-  if (thermalPadObstacles.length === 0) return true
-  if (thermalPadObstacles.length > MAX_QFP_THERMAL_PAD_COUNT) return false
-
-  const perimeterBounds = getBoundsForObstacles(perimeterPadObstacles)
-  const width = perimeterBounds.maxX - perimeterBounds.minX
-  const height = perimeterBounds.maxY - perimeterBounds.minY
-  const centralMinX = perimeterBounds.minX + width * 0.2
-  const centralMaxX = perimeterBounds.maxX - width * 0.2
-  const centralMinY = perimeterBounds.minY + height * 0.2
-  const centralMaxY = perimeterBounds.maxY - height * 0.2
-  const perimeterPadThickness = Math.min(
-    ...perimeterPadObstacles.map((obstacle) =>
-      Math.min(obstacle.width, obstacle.height),
-    ),
-  )
-
-  return thermalPadObstacles.every((obstacle) => {
-    const minSide = Math.min(obstacle.width, obstacle.height)
-
-    return (
-      obstacle.center.x >= centralMinX &&
-      obstacle.center.x <= centralMaxX &&
-      obstacle.center.y >= centralMinY &&
-      obstacle.center.y <= centralMaxY &&
-      minSide > perimeterPadThickness
-    )
-  })
-}
-
-function hasQfpPerimeterPadRing(perimeterPadObstacles: Obstacle[]) {
-  if (perimeterPadObstacles.length < MIN_QFP_PADS_PER_SIDE * 4) return false
-  if (perimeterPadObstacles.length > MAX_QFP_PAD_COUNT) return false
-
-  const { counts, maxNearestSideRatio } = getNearestSideCounts(
-    perimeterPadObstacles,
-  )
-
-  return (
-    counts.top >= MIN_QFP_PADS_PER_SIDE &&
-    counts.right >= MIN_QFP_PADS_PER_SIDE &&
-    counts.bottom >= MIN_QFP_PADS_PER_SIDE &&
-    counts.left >= MIN_QFP_PADS_PER_SIDE &&
-    maxNearestSideRatio <= MAX_QFP_CENTER_NEAREST_SIDE_RATIO
-  )
-}
-
-function splitQfpMemberObstacles(memberObstacles: Obstacle[]) {
-  const perimeterPadObstacles = getQfpPerimeterPadObstacles(memberObstacles)
-  const thermalPadObstacles = memberObstacles.filter(
-    (obstacle) => !isQfpPerimeterPadObstacle(obstacle),
-  )
-
-  return { perimeterPadObstacles, thermalPadObstacles }
-}
-
-function isQfpThermalPadLikeComponent(memberObstacles: Obstacle[]) {
-  const { perimeterPadObstacles, thermalPadObstacles } =
-    splitQfpMemberObstacles(memberObstacles)
-
-  if (thermalPadObstacles.length === 0) return false
-  if (!hasQfpPerimeterPadRing(perimeterPadObstacles)) return false
-
-  return areCentralThermalPadObstacles({
-    perimeterPadObstacles,
-    thermalPadObstacles,
-  })
-}
-
-function isQfpLikeComponent(memberObstacles: Obstacle[]) {
-  const { perimeterPadObstacles, thermalPadObstacles } =
-    splitQfpMemberObstacles(memberObstacles)
-
-  if (thermalPadObstacles.length > 0) return false
-  if (
-    !areCentralThermalPadObstacles({
-      perimeterPadObstacles,
-      thermalPadObstacles,
-    })
-  ) {
-    return false
-  }
-
-  return hasQfpPerimeterPadRing(perimeterPadObstacles)
-}
-
-function detectComponentKind(
-  memberObstacles: Obstacle[],
-  inputSrj: SimpleRouteJson,
-): ComponentKind | null {
-  if (isQfpThermalPadLikeComponent(memberObstacles)) return "qfp_thermalpad"
-  if (isQfpLikeComponent(memberObstacles)) return "qfp"
-  if (isSoicLikeComponent({ memberObstacles, inputSrj })) return "soic"
-  if (isBgaLikeComponent(memberObstacles)) return "bga"
-
-  return null
-}
 
 /**
  * Current detection stage: groups SRJ obstacles by component and replaces each
@@ -666,10 +317,10 @@ export class RectBoundsComponentDetectionStage extends BaseSolver {
 
     const detectedEntries = Object.entries(grouped).filter(
       ([componentId, memberObstacles]) => {
-        const componentKind = detectComponentKind(
+        const componentKind = detectComponentKind({
           memberObstacles,
-          this.inputSrj,
-        )
+          inputSrj: this.inputSrj,
+        })
         if (!componentKind) return false
 
         componentKinds[componentId] = componentKind
